requirements.txt
alembic==0.9.3
autopep8==1.3.2
click==6.7
flake8==3.3.0
Flask==0.12.2
Flask-Migrate==2.0.4
Flask-Redis==0.3.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pycrypto==2.6.1
pyflakes==1.5.0
PyMySQL==0.7.11
python-dateutil==2.6.0
python-editor==1.0.3
redis==2.10.5
six==1.10.0
SQLAlchemy==1.1.11
Werkzeug==0.12.2
